refactor: 라우터 타입 분리 및 로직 최적화

+++ backend/src/router/auth.ts
@@ -1,27 +1,9 @@
-import express, { Request, Response, Router } from "express";
-import { PrismaClient, UserSession } from "@prisma/client";
+import express, { Router } from "express";
+import { PrismaClient, User, UserSession, UserTendency } from "@prisma/client";
+import { AuthGetRequest, AuthPostRequest } from "./types/auth";
 
-interface UserInfo {
-  id: string;
-  name: string;
-  phone: number;
-  email: string;
-  createdAt: Date;
-  updatedAt: Date | null;
-}
-
-interface AuthGetRequest extends Request {
-  readonly params: {
-    id: string;
-  }
-}
-
-interface AuthPostRequest extends Request {
-  readonly body: {
-    id: string;
-    pw?: string;
-    sessionId?: string
-  }
+interface UserInfo extends User {
+  tendency: UserTendency | null
 }
 
 const router: Router = express.Router();
@@ -30,69 +12,58 @@ const prisma = new PrismaClient();
 /**
  * 세션 생성
  */
-router.post('/', (req: AuthPostRequest, res: Response) => {
-    if(req.body.pw !== undefined && req.ip) {
-    const ip = req.ip === '::1' ? '127.0.0.1' : req.ip;
-    prisma.user.findUniqueOrThrow({
-      select: {
-        id: true,
-        name: true,
-        phone: true,
-        email: true,
-        createdAt: true,
-        updatedAt: true
-      },
-      where: {
-        id: req.body.id,
-        pw: req.body.pw
-      }
-    }).then((user: UserInfo) => {
-      prisma.userSession.deleteMany({
+router.post('/', async (req: AuthPostRequest, res: ExpressResponse) => {
+  try {
+    if(req.body.pw !== undefined && req.ip !== undefined) {
+      const ip = req.ip === '::1' ? '127.0.0.1' : req.ip;
+      const user: UserInfo = await prisma.user.findUniqueOrThrow({
+        include: {
+          tendency: true
+        },
         where: {
-          userId: req.body.id
+          id: req.body.id,
+          pw: req.body.pw
         }
-      }).then(() => {
-        prisma.userSession.create({
-          select: {
-            id: true,
-          },
-          data: {
-            ip: ip,
-            userId: req.body.id
-          }
-        }).then(async (session) => {
-          await prisma.userLoginHistory.create({
-            data: {
-              userId: req.body.id,
-              ip: ip,
-              success: true
-            }
-          })
-          res.status(200).json({ session: session.id, user: user });
-        }).catch(async () => {
-          await prisma.userLoginHistory.create({
-            data: {
-              userId: req.body.id,
-              ip: ip,
-              success: false
-            }
-          })
-          res.status(500).send("Internal Server Error")
-        })
-      }).catch(() => {
-        res.status(500).send("Internal Server Error")
+      });
+
+      await prisma.userSession.deleteMany({
+        where: {
+          userId: user.id
+        }
       })
-    }).catch(async () => {
+        
       await prisma.userLoginHistory.create({
         data: {
           userId: req.body.id,
           ip: ip,
+          success: true
+        }
+      })
+
+      const sessionId = (await prisma.userSession.create({
+        select: {
+          id: true,
+        },
+        data: {
+          ip: ip,
+          userId: req.body.id
+        }
+      })).id
+        
+      res.status(200).json({ session: sessionId, user: user });
+    } else {
+      new Error("세션 생성 인자 부족");
+    }
+  } catch {
+    if(req.body.id !== undefined && req.ip !== undefined) {
+      await prisma.userLoginHistory.create({
+        data: {
+          userId: req.body.id,
+          ip: req.ip === '::1' ? '127.0.0.1' : req.ip,
           success: false
         }
       })
-      res.status(500).send("Internal Server Error")
-    })
-  } else {
+    }
     res.status(500).send("Internal Server Error")
   }
 })
@@ -100,55 +71,56 @@ router.post('/', (req: AuthPostRequest, res: Response) => {
 /**
  * 세션 검색
  */
-router.get('/:id', (req: AuthGetRequest, res: Response) => {
-  const ip = req.ip === '::1' ? '127.0.0.1' : req.ip;
-  prisma.userSession.findFirstOrThrow({
-    select: {
-      id: true,
-      userId: true,
-    },
-    where: {
-      id: req.params.id,
-      ip: ip
-    }
-  }).then((session) => {
-    prisma.user.findUniqueOrThrow({
-      select: {
-        id: true,
-        name: true,
-        phone: true,
-        email: true,
-        createdAt: true,
-        updatedAt: true
-      },
-      where: {
-        id: session.userId
-      }
-    }).then((user: UserInfo) => {
+router.get('/:id', async (req: AuthGetRequest, res: ExpressResponse) => {
+  try {
+    if(req.ip !== undefined) {
+      const ip = req.ip === '::1' ? '127.0.0.1' : req.ip;
+      const session: UserSession = await prisma.userSession.findFirstOrThrow({
+        where: {
+          id: req.params.id,
+          ip: ip
+        }
+      });
+      const user: UserInfo = await prisma.user.findUniqueOrThrow({
+        include: {
+          tendency: true
+        },
+        where: {
+          id: session.userId
+        }
+      });
+
       res.status(200).json({ session: session.id, user: user });
-    }).catch(() => {
-      res.status(500).send("Internal Server Error")
-    })
-  }).catch(() => {
+    } else {
+      new Error("세션 검색 인자 부족");
+    }
+  } catch {
     res.status(500).send("Internal Server Error")
-  })
+  }
 })
 
 /**
  * 세션 제거
  */
-router.post('/delete', (req: AuthPostRequest, res: Response) => {
-  const ip = req.ip === '::1' ? '127.0.0.1' : req.ip;
-  prisma.userSession.delete({
-    where: {
-      id: req.body.id,
-      ip: ip
+router.post('/delete', async (req: AuthPostRequest, res: ExpressResponse) => {
+  try {
+    if(req.ip !== undefined) {
+      const ip = req.ip === '::1' ? '127.0.0.1' : req.ip;
+
+      await prisma.userSession.delete({
+        where: {
+          id: req.body.id,
+          ip: ip
+        }
+      })
+
+      res.status(200).json({ message: "session deleted" });
+    } else {
+      new Error("세션 제거 인자 부족");
     }
-  }).then(() => {
-    res.status(200).json({ message: "session deleted" });
-  }).catch(() => {
+  } catch {
     res.status(500).send("Internal Server Error")
-  })
+  }
 })
 
 export default router;
